+++ AddOns/HandyNotes_Oribos/ChangeLog.md
@@ -1,5 +1,16 @@
 ## Revision History:
 
+#### v13 (July 04, 2020)
+-------------------------------
+* Added the new "Portal" to Keeper's Respite (Korthia) [Github #5](https://github.com/Dathwada/handynotes-oribos/issues/5)
+* Fixed portal icon behavior (the icon is now red as intended if the required quest isn't completed)
+* Updated localizations (enUS, deDE, ptBr, ruRU)
+
+#### v12 (June 30, 2020)
+-------------------------------
+* Updated localizations
+* Bump .toc to 90100
+
 #### v11 (March 10, 2020)
 -------------------------------
 * Added Korean localizationby [unrealcrom96](https://www.curseforge.com/members/unrealcrom96)